+++ cppRFM/implement/DeepBSDE/src/config.cpp
@@ -24,10 +24,10 @@ Config load_config(const std::string& json_path)
 
     NetConfig net_cfg =
     {
-        j["net_config"]["y_init_range"].get<std::vector<double>>(),
-        j["net_config"]["num_hiddens"].get<std::vector<int>>(),
-        j["net_config"]["lr_values"].get<std::vector<double>>(),
-        j["net_config"]["lr_boundaries"].get<std::vector<int>>(),
+        j["net_config"]["y_init_range"].get<std::vector<float>>(),
+        j["net_config"]["num_hiddens"].get<std::vector<int64_t>>(),
+        j["net_config"]["lr_values"].get<std::vector<float>>(),
+        j["net_config"]["lr_boundaries"].get<std::vector<int64_t>>(),
         j["net_config"]["num_iterations"],
         j["net_config"]["batch_size"],
         j["net_config"]["valid_size"],
