+++ cppRFM/implement/DeepBSDE/src/mlp.cpp
@@ -2,13 +2,13 @@
 
 MLPImpl::MLPImpl(const Config& config)
 {
-    int input_dim = config.eqn_config.dim;
-    int output_dim = config.eqn_config.dim;
-	std::vector<int> num_hiddens = config.net_config.num_hiddens;
+    int64_t input_dim = config.eqn_config.dim;
+    int64_t output_dim = config.eqn_config.dim;
+	std::vector<int64_t> num_hiddens = config.net_config.num_hiddens;
     bn_layers_.push_back(register_module("bn_input", torch::nn::BatchNorm1d(input_dim)));
     for (size_t i = 0; i < num_hiddens.size(); ++i)
     {
-        int hidden = num_hiddens[i];
+        int64_t hidden = num_hiddens[i];
 
         auto linear = torch::nn::Linear(torch::nn::LinearOptions(input_dim, hidden).bias(false));
         dense_layers_.push_back(register_module("dense_" + std::to_string(i), linear));
@@ -26,7 +26,7 @@ MLPImpl::MLPImpl(const Config& config)
 torch::Tensor MLPImpl::forward(torch::Tensor x)
 {
     x = bn_layers_[0]->forward(x);
-    for (size_t i = 0; i < dense_layers_.size() - 1; ++i)
+    for (int64_t i = 0; i < dense_layers_.size() - 1; ++i)
     {
         x = dense_layers_[i]->forward(x);
         x = bn_layers_[i + 1]->forward(x);
